pyproject.toml
[project]
authors = [
    {name = "ADrone7"},
]
name = "hexlet-code"
version = "0.1.1"
description = "Compare two json or yaml files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]

[tool.coverage.run]
disable_warnings = ["no-data-collected"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
